requirements.txt
feedparser>=6.0
jinja2>=3.1
google-genai>=1.0
anthropic>=0.40
gspread>=6.0
google-auth>=2.0
premailer>=3.10
python-dotenv>=1.0
pyyaml>=6.0
pytest>=8.0
